feat(word): 更新Word互操作接口并优化实现
- 为IWordHyperlinks接口添加ComCollectionWrap特性
- 将IWordHyperlinks的Application属性标记为可空并添加ComPropertyWrap特性
- 为IWordHyperlinks添加索引器重载支持字符串索引
- 优化IWordHyperlinks.Add方法参数为可空类型并设置默认值
- 为IWordKeyBinding接口添加ComObjectWrap特性
- 将IWordKeyBinding的Application和Parent属性标记为可空
- 添加Protected属性的文档注释

refactor(word): 移除内部实现类并更新依赖版本

- 删除WordHyperlinks和WordKeyBinding内部实现类
- 更新WordSamples和ExcelSamples中的包版本从1.2.2到2.0.1
- 切换项目引用方式从ProjectReference到PackageReference

fix(word): 修复文档操作和图形处理中的问题

- 使用document.SaveAs替代document.Save方法
- 将颜色值处理从unchecked转换为System.Drawing.Color
- 更新SmartArt文本框访问方式从TextFrame到TextFrame2
- 移除文档统计分析相关代码

+++ Samples/WordSamples/GraphicsAndImageOperationsSample/GraphicEffectsHelper.cs
@@ -139,7 +139,7 @@ namespace GraphicsAndImageOperationsSample
                 }
 
                 // 设置阴影属性
-                shape.Shadow.ForeColor.RGB = (int)color;
+                shape.Shadow.ForeColor.RGB = color;
                 shape.Shadow.OffsetX = offsetX;
                 shape.Shadow.OffsetY = offsetY;
                 shape.Shadow.Blur = blur;
